LayoutTests/http/tests/preload/download_resources_from_header_iframe-expected.txt

CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: <link rel=preload> cannot have the empty string as `as` value


--------
Frame: '<!--frame1-->'
--------
PASS internals.isPreloaded('../resources/dummy.js'); is true
PASS internals.isPreloaded('../resources/dummy.css'); is true
PASS internals.isPreloaded('../resources/square.png'); is true
PASS internals.isPreloaded('../resources/Ahem.ttf'); is true
PASS internals.isPreloaded('../resources/test.mp4'); is true
PASS internals.isPreloaded('../resources/test.oga'); is true
PASS internals.isPreloaded('../security/resources/captions.vtt'); is true
PASS internals.isPreloaded('../resources/dummy.xml?badvalue'); is false
PASS internals.isPreloaded('../resources/dummy.xml?empty'); is false
PASS internals.isPreloaded('../resources/dummy.xml'); is true
PASS successfullyParsed is true

TEST COMPLETE
